requirements.txt
# Card Print Generator Dependencies

# Image processing library
Pillow>=10.0.0
